refactor: add <GameBoard />
